refactor: use laravel-data for AutomaticTorrentFreeleech form requests
